fix build when no RSA (or no curve) is selected
AMCL should support compiling with no RSA level selected or no curve
selected, to support users that are using one or the other.  This
commit adds such support to the build.

Two changes were needed.

First, the AMCL_RSA and AMCL_CURVE must be quoted when passed to
"REGEX REPLACE", in case they are empty.

Second, the "src/rsa_support.c" and "src/ecdh_support.c" sources
should only be copied to the build tree if an RSA level or a curve was
selected.